refactor: 페이지 조회 path 경로 매칭 파일 분리
