+++ src/components/commons/track/Tracking.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TRACK_EVENTS } from "src/track/events";
+import { TRACK_EVENTS } from "src/track/constants/events";
 import { trackEvent } from "src/track/track";
 
 type TrackingProps = {
