+++ src/track/useTrackingPageView.ts
@@ -2,30 +2,29 @@ import mixpanel from "mixpanel-browser";
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 import { trackEvent } from "./track";
-import { TRACK_EVENTS } from "./events";
+import { TRACK_EVENTS } from "./constants/events";
+import { extractIdFromPath } from "./extractIdFromPath";
+import { TRACKING_CONFIG } from "./constants/trackingConfig";
 
 export function useTrackingPageView(): void {
   const location = useLocation();
 
   useEffect(() => {
     if (mixpanel && typeof mixpanel.track === "function") {
-      const path = location.pathname;
+      const { pathname } = location;
 
-      if (path.startsWith("/gig")) {
-        const match = path.match(/^\/gig\/(\w+)/);
-        const gigId = match ? match[1] : undefined;
-        trackEvent(TRACK_EVENTS.VIEWED_PAGE_GIG, gigId ? { gig_id: gigId } : undefined);
+      const matchedConfig = TRACKING_CONFIG.find((config) => pathname.startsWith(config.path));
+
+      if (matchedConfig) {
+        const paramValue = extractIdFromPath(pathname, matchedConfig.path);
+        trackEvent(
+          TRACK_EVENTS[matchedConfig.event],
+          paramValue ? { [matchedConfig.paramKey]: paramValue } : undefined
+        );
         return;
       }
 
-      if (path.startsWith("/book")) {
-        const match = path.match(/^\/book\/(\w+)/);
-        const gig_id = match ? match[1] : undefined;
-        trackEvent(TRACK_EVENTS.VIEWED_PAGE_BOOK, gig_id ? { gig_id: gig_id } : undefined);
-        return;
-      }
-
-      trackEvent(TRACK_EVENTS.PAGE_VIEW, { path });
+      trackEvent(TRACK_EVENTS.PAGE_VIEW, { path: pathname });
     }
   }, [location.pathname]);
 }
